Restructure to use vertical bar chart
