+++ blog/tests/factories.py
@@ -61,7 +61,6 @@ class BlogPageFactory(wagtail_factories.PageFactory):
                     'styled_text',
                     'vertical_bar_chart',
                     'tree_map_chart',
-                    'stacked_bar_chart',
                     'bubble_map_chart',
                 ],
             )
@@ -84,7 +83,6 @@ class BlogPageFactory(wagtail_factories.PageFactory):
             'styled_text',
             'vertical_bar_chart',
             'tree_map_chart',
-            'stacked_bar_chart',
             'bubble_map_chart',
         ],
     )
