+++ common/templates/common/_video_or_image.html
@@ -12,8 +12,6 @@
 					{% include 'common/blocks/vertical_bar_chart.html' with value=image.0.value static_graphic=true %}
 				{% elif image.0.block_type == "tree_map_chart" %}
 					{% include 'common/blocks/tree_map_chart.html' with value=image.0.value static_graphic=true %}
-				{% elif image.0.block_type == "stacked_bar_chart" %}
-					{% include 'common/blocks/stacked_bar_Chart.html' with value=image.0.value static_graphic=true %}
 				{% elif image.0.block_type == "bubble_map_chart" %}
 					<div class="primary-banner-media__inner">
 						{% include 'common/blocks/bubble_map_chart.html' with value=image.0.value static_graphic=true %}
