--- common/templates/common/blocks/stacked_bar_chart.html
@@ -1,39 +1,0 @@
-{% if static_graphic is None and value.title %}<h2>{{ value.title }}</h2>{% endif %}
-<div class="chartContainer">
-	<div
-		class="chart-stacked-bar"
-		{% if value.incident_set.categories %}
-			{% for category in value.incident_set.categories %}
-				data-category-{{ forloop.counter }}="{{ category }}"
-			{% endfor %}
-		{% endif %}
-		{% if value.incident_set.tag %}
-			data-tag="{{ value.incident_set.tag }}"
-		{% endif %}
-		{% if value.incident_set.lower_date %}
-			data-lower-date="{{ value.incident_set.lower_date|date:'c' }}"
-		{% endif %}
-		{% if value.incident_set.upper_date %}
-			data-upper-date="{{ value.incident_set.upper_date|date:'c' }}"
-		{% endif %}
-		{% if value.description %}
-			data-description="{{ value.description }}"
-		{% endif %}
-		{% if value.title %}
-			data-title="{{ value.title }}"
-		{% endif %}
-		{% if static_graphic is None %}
-			data-interactive="true"
-		{% endif %}
-		{% if mini_graphic is None %}
-			data-full-size="true"
-		{% endif %}
-		data-branch-field-name="{{ value.branch_field_name }}"
-		data-branches="{{ value.branches }}"
-		data-url="{{ value.data_url }}"
-	>
-		{% if mini_graphic is not None %}
-			<svg viewBox="0 0 655 440" width="100%" style="display: block; background-color: #fafafa" />
-		{% endif %}
-	</div>
-</div>
