+++ common/tests/utils.py
@@ -196,23 +196,6 @@ def generate_tree_map_chart():
     )
 
 
-def generate_stacked_bar_chart():
-    incident_set = {
-        'categories': [],
-        'tag': None,
-        'lower_date': fake.date_between('-2y', '-1y'),
-        'upper_date': fake.date_between('-11M'),
-    }
-    return generate_field(
-        'stacked_bar_chart',
-        {
-            'title': make_words().capitalize(),
-            'incident_set': incident_set,
-            'description': fake.text(max_nb_chars=200),
-        }
-    )
-
-
 def generate_bubble_map_chart():
     incident_set = {
         'categories': [],
@@ -396,7 +379,6 @@ class StreamfieldProvider(BaseProvider):
             'aside': generate_aside,
             'vertical_bar_chart': generate_vertical_bar_chart,
             'tree_map_chart': generate_tree_map_chart,
-            'stacked_bar_chart': generate_stacked_bar_chart,
             'bubble_map_chart': generate_bubble_map_chart,
         }
